// cli/src/generated/control_panel.d.ts
export interface AddRegistryEntryResponse { 'entry' : RegistryEntry }
export type AddRegistryEntryResult = { 'Ok' : AddRegistryEntryResponse } |
  { 'Err' : ApiError };
export interface ApiError {
  'code' : string,
  'message' : [] | [string],
  'details' : [] | [Array<[string, string]>],
}
// ...
export interface EditRegistryEntryResponse { 'entry' : RegistryEntry }
export type EditRegistryEntryResult = { 'Ok' : EditRegistryEntryResponse } |
  { 'Err' : ApiError };
// ...
export interface RegistryEntry {
  'id' : UUID,
  'categories' : Array<string>,
  'updated_at' : [] | [TimestampRFC3339],
  'value' : RegistryEntryValue,
  'metadata' : Array<Metadata>,
  'name' : string,
  'tags' : Array<string>,
  'description' : string,
  'created_at' : TimestampRFC3339,
}